Use both IPv4+IPv6 sockets to check whether a port is free.
This allows the portpicker to function in dual-stack and IPv4/IPv6-only
environments.

For a port to be considered available, the kernel must support at least
one of (IPv6, IPv4), and the port must be available on each supported
family.

Use identical/redundant implementations of bind() and is_port_free()
between portpicker and portserver.  I am assuming that these were
intentionally designed as self-contained programs, instead of sharing a
common library.