Make portpicker_test more robust, especially when run on a busy CI host. For (#15)
tests that use portpiker without server, don't fail on the first failure to
grab a port; try a few times. For testIsPortFree, don't fail outright when a
port is stolen after portpicker guessed it was free, but try a few times. As
a general measure, expecting >95% success rate seems rasonable.

Rewrite testPickPortsWithError, which appeared to be trying to test the two
parts of _pick_unused_port_without_server by playing chance games (that
failed ~10-15% of the time), so that it explicitly covers both parts by
mocking out the right function. The actual picking of the port can still
fail, since portpicker without server is inherently flaky, so try a few
times before giving up.

With these changes, there were no failures in 10000 runs of the test (many
in parallel).