Allow the CLI to keep a port bound after exit.
Useful for non-portserver cases as a hack to avoid race conditions.

This logic with the fork(), stdio close, and sleep to hold the bind
comes from our internal `googletest.sh` bash test framework.  Putting it
here gets rid of a couple `perl -e` and `python -c` uses within that
framework.